% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMultiomics.R
\name{BayesMultiomics.CV}
\alias{BayesMultiomics.CV}
\title{BayesMultiomics.CV}
\usage{
BayesMultiomics.CV(
  M,
  G,
  grouping,
  nu0 = 0.5,
  nu1 = 10^3,
  nu = 1,
  lambda = 1,
  a = 1,
  b = 1,
  EMVS_I = 10,
  EMVS_thresh = 1e-04,
  Y,
  C,
  Delta,
  a0,
  gstr,
  n_fold = 10,
  random_seed = NULL,
  NEG_I = 10,
  NEG_thresh = 1e-04,
  lower_R2 = 0.2,
  upper_R2 = 0.8,
  transform_M = "L",
  .mpmath = setup_mpmath()
)
}
\arguments{
\item{M}{DDNA Methylation matrix, a matrix with rows corresponding to samples and columns corresponding to probes/sites.}

\item{G}{Gene expression level, a matrix with rows corresponding to samples and columns corresponding to genes.}

\item{grouping}{Gene grouping, a vector of group membership of each gene. This could be given by the user or obatined by performing DAVID Functional Classification of genes. See vignette("gene_grouping").}

\item{nu0}{Parameter \code{nu0} for spike-and-slab Gaussian mixture prior on \eqn{\omega} (orginally \eqn{\beta} in Veronika Ročková & Edward I. George (2013)).}

\item{nu1}{parameter \code{nu1} for spike-and-slab Gaussian mixture prior on \eqn{\omega} (orginally \eqn{\beta} in Veronika Ročková & Edward I. George (2013)).}

\item{nu}{For the prior on \eqn{\sigma^2}, an inverse gamma prior \eqn{\pi(\sigma2 | \gamma) = IG(\nu/2, \nu\lambda/2)}.}

\item{lambda}{For the prior on \eqn{\sigma^2}, an inverse gamma prior \eqn{\pi(\sigma2 | \gamma) = IG(\nu/2, \nu\lambda/2)}.}

\item{a}{Parameter \eqn{a} for the beta prior \eqn{\pi(\theta) \propto \theta a-1(1-\theta)b-1}, \eqn{a, b > 0}.}

\item{b}{Parameter \eqn{b} for the beta prior \eqn{\pi(\theta) \propto \theta a-1(1-\theta)b-1}, \eqn{a, b > 0}.}

\item{EMVS_I}{Maximum number of iterations of EMVS.}

\item{EMVS_thresh}{Threshold for convergence criterion for EMVS.}

\item{Y}{Clinical outcome, a vector with each element corresponding to the clinical outcome of a sample.}

\item{C}{Clinical feature, a matrix with rows corresponding to samples and columns corresponding to clinical features.}

\item{Delta}{Delta.}

\item{a0}{Shape parameter of the Gamma distribution prior. This parameter affects the within-group and overall sparsity.}

\item{gstr}{String. Prior with two options: "scale", for applying rescaling, or "1", for not.}

\item{n_fold}{Number of folds in K-fold cross validation.}

\item{random_seed}{Random seed for splitting folds for cross validation.}

\item{NEG_I}{Number of maximum iterations for NEG_em.}

\item{NEG_thresh}{Convergence criterion fpr NEG_em.}

\item{lower_R2}{Lower limit for R2 threshold, default = 0.2.}

\item{upper_R2}{Upper limit for R2 threshold, default = 0.8.}

\item{transform_M}{Transform methylation matrix \code{M}. Options: "L" = linear transformation, "Q" = quadratic transformation; "QS" = cubic spline transformation.}

\item{.mpmath}{(optional) function depends on mpmath package from python. pointer for mpmath package}
}
\description{
Run first and second stage models in \emph{Hao Xue, Sounak Chakraborty & Tanujit Dey (2024), Bayesian Shrinkage Models for Integration and Analysis of Multiplatform High-dimensional Genomics Data}  and \emph{Mansoo Cho, Tanujit Dey, Hao Xue & Sounak Chakraborty (2025), BayesMultiomics: An R Package for Bayesian Shrinkage Models for Integration and Analysis of Multi-platform High-dimensional Genomics Data}
In the first stage, Expectation–Maximization Variable Selection (EMVS) is used to learn the regulating mechanism between epigenomics (e.g., gene methylation) and gene expression while considering functional gene annotations.
Genes are grouped based on the regulating mechanism learned in the first stage using the function, EMVS.
Then, a group-wise penalty is applied to select genes significantly associated with clinical outcomes while incorporating clinical features using Bayesian variable selection model with extended NEG priors in \emph{Veronika Rockova & Edward I. George (2013): EMVS: The EM Approach to Bayesian Variable Selection, Journal of the American Statistical Association}.
K-fold Cross validation is performed to evaluate model performances. The coefficients from all models are averaged.
}
\details{
\code{gstr} will take two options "scale" or "1." if \code{gstr} == "scale" then g = 1/N^2 where N = number of genes
}
\examples{
# params
M <- GBM_data2$M
G <- GBM_data2$G
# fp <- system.file("eg_fx_classification.txt", package="EMMultiOmics")
fp <- system.file("eg_fx_classification.txt", package="BayesMultiomics")
gene_grouping <- DAVIDGeneGrouping(eg_gene_symbols, fp)
Y <- GBM_data2$Y
C <- GBM_data2$C
a0 <- 0.1
gstr <- "scale"
Delta <- GBM_data2$Delta

# run
multiomics_cv <- BayesMultiomics.CV(
  M=M,
  G=G,
  grouping=gene_grouping,
  Y=Y,
  C=C,
  a0=a0,
  gstr=gstr,
  Delta=Delta,
  n_fold=10,
  random_seed = 123,
  EMVS_I = 10,
  NEG_I=20,
  EMVS_thresh = 0.0001,
  NEG_thresh = 0.0001
)

# plot
plot(multiomics_cv)

}
