% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMultiomics.R
\name{BayesMultiomics}
\alias{BayesMultiomics}
\title{BayesMultiomics}
\usage{
BayesMultiomics(
  M,
  G,
  grouping,
  nu0 = 0.5,
  nu1 = 10^3,
  nu = 1,
  lambda = 1,
  a = 1,
  b = 1,
  EMVS_I = 10,
  EMVS_thresh = 1e-04,
  transform_M = "L",
  Y,
  C,
  a0,
  gstr,
  NEG_I = 10,
  NEG_thresh = 1e-04,
  lower_R2 = 0.2,
  upper_R2 = 0.8,
  .mpmath = setup_mpmath()
)
}
\arguments{
\item{M}{DNA Methylation matrix, a matrix with rows corresponding to samples and columns corresponding to probes/sites.}

\item{G}{gene expression level}

\item{grouping}{Gene grouping, a vector of group membership of each gene. This could be given by the user or obatined by performing DAVID Functional Classification of genes. See vignette("gene_grouping").}

\item{nu0}{Parameter \code{nu0} for spike-and-slab Gaussian mixture prior on \eqn{\beta}}

\item{nu1}{parameter \code{nu1} for spike-and-slab Gaussian mixture prior on \eqn{\beta}}

\item{lambda}{For the prior on \eqn{\sigma2}, an inverse gamma prior \eqn{\pi(\sigma2 | \gamma) = IG(\nu/2, \nu\lambda/2)}}

\item{a}{Parameter \eqn{a} for the beta prior \eqn{\pi(\theta) \propto \theta a-1(1-\theta)b-1}, \eqn{a, b > 0}}

\item{b}{Parameter \eqn{b} for the beta prior \eqn{\pi(\theta) \propto \theta a-1(1-\theta)b-1}, \eqn{a, b > 0}}

\item{EMVS_I}{Maximum number of iterations of EMVS}

\item{EMVS_thresh}{Threshold for convergence criterion for EMVS}

\item{transform_M}{transform methylation matrix \code{M}. Options: "L" = linear transformation, "Q" = quadratic transformation; "QS" = cubic spline transformation}

\item{Y}{clinical outcome}

\item{C}{clinical features}

\item{a0}{size sparsity}

\item{gstr}{prior, two options: "scale" or "1"}

\item{NEG_I}{number of maximum iterations for NEG_em}

\item{NEG_thresh}{convergence criterion fpr NEG_em}

\item{lower_R2}{lower limit for R2 threshold, default = 0.2}

\item{upper_R2}{upper limit for R2 threshold, default = 0.8}

\item{.mpmath}{(Optional) function depends on mpmath package from python. pointer for mpmath package. Default: setup_mpmath()}
}
\description{
Run first and second stage models in \emph{Hao Xue, Sounak Chakraborty & Tanujit Dey (2024), Bayesian Shrinkage Models for Integration and Analysis of Multiplatform High-dimensional Genomics Data}.
In the first stage, Expectation–Maximization Variable Selection (EMVS) is used to learn the regulating mechanism between epigenomics (e.g., gene methylation) and gene expression while considering functional gene annotations.
Genes are grouped based on the regulating mechanism learned in the first stage using the function, EMVS.
Then, a group-wise penalty is applied to select genes significantly associated with clinical outcomes while incorporating clinical features using Bayesian variable selection model with extended NEG priors in \emph{Veronika Rockova & Edward I. George (2013): EMVS: The EM Approach to Bayesian Variable Selection, Journal of the American Statistical Association}.
}
\details{
\code{gstr} will take two options "scale" or "1." if \code{gstr} == "scale" then g = 1/N^2 where N = number of genes
}
\examples{
# params
M <- GBM_data2$M
G <- GBM_data2$G
fp <- system.file("eg_fx_classification.txt", package="EMMultiOmics")
# fp <- system.file("eg_fx_classification.txt", package="BayesMultiomics")
gene_grouping <- DAVIDGeneGrouping(eg_gene_symbols, fp)
Y <- GBM_data2$Y
C <- GBM_data2$C
a0 <- 0.1
gstr <- "scale"
mpmath <- setup_mpmath()

# run
multiomics <- BayesMultiomics(
  M=M,
  G=G,
  grouping=gene_grouping,
  Y=Y,
  C=C,
  a0=a0,
  gstr=gstr,
  EMVS_I = 10,
  NEG_I= 10,
  EMVS_thresh = 0.0001,
  NEG_thresh = 0.0001,
  transform_M = "L",
  lower_R2 = 0.2,
  upper_R2 = 0.8,
  .mpmath = mpmath
)

}
