% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMultiomics.R
\name{BayesMultiomicsPriorSensitivity}
\alias{BayesMultiomicsPriorSensitivity}
\title{BayesMultiomicsPriorSensitivity}
\usage{
BayesMultiomicsPriorSensitivity(
  gstr_vec,
  a0_vec,
  M,
  G,
  grouping,
  nu0 = 0.5,
  nu1 = 10^3,
  nu = 1,
  lambda = 1,
  a = 1,
  b = 1,
  EMVS_I = 10,
  EMVS_thresh = 1e-04,
  Y,
  C,
  Delta,
  n_fold = 10,
  random_seed = NULL,
  NEG_I = 10,
  NEG_thresh = 1e-04,
  lower_R2 = 0.2,
  upper_R2 = 0.8,
  transform_M = "L",
  .mpmath = setup_mpmath()
)
}
\arguments{
\item{gstr_vec}{a vector of \code{g} values, two options: "scale" or 1}

\item{a0_vec}{a vector of \code{a} values}

\item{...}{Additional parameters for \code{BayesMultiomics}}
}
\description{
Run \code{BayesMultiomics} for given sets of parameters \code{g} and \code{a}
}
\examples{
# params
M <- GBM_data2$M
G <- GBM_data2$G
# fp <- system.file("eg_fx_classification.txt", package="EMMultiOmics")
fp <- system.file("eg_fx_classification.txt", package="BayesMultiomics")
gene_grouping <- DAVIDGeneGrouping(eg_gene_symbols, fp)
Y <- GBM_data2$Y
C <- GBM_data2$C
Delta <- GBM_data2$Delta

a0 <- c(0.1, 1, 10, 50)
g <- list("scale", 1)


# run
multiomics_sensitivity <- BayesMultiomicsPriorSensitivity(
  M=M,
  G=G,
  grouping=gene_grouping,
  Y=Y,
  C=C,
  a0_vec=a0,
  gstr_vec=g,
  Delta=Delta,
  n_fold=10,
  random_seed = 123,
  EMVS_I = 10,
  NEG_I=20,
  EMVS_thresh = 0.0001,
  NEG_thresh = 0.0001
)

}
