% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSDIM.R
\name{.NEG_em}
\alias{.NEG_em}
\title{Normal-Exponential-Gamma EM Algorithm}
\usage{
.NEG_em(
  Y,
  G,
  C,
  a0,
  gstr,
  Zmatrix,
  I = 10,
  thresh = 0.001,
  .mpmath = setup_mpmath()
)
}
\arguments{
\item{Y}{Clinical outcome, a vector with each element corresponding to the clinical outcome of a sample.}

\item{G}{Gene expression level, a matrix with rows corresponding to samples and columns corresponding to genes.}

\item{C}{Clinical feature, a matrix with rows corresponding to samples and columns corresponding to clinical features.}

\item{a0}{Shape parameter of the Gamma distribution prior. This parameter affects the within-group and overall sparsity.}

\item{gstr}{String. Prior with two options: "scale", for applying rescaling, or "1", for not.}

\item{Zmatrix}{Loading matrix, a matrix with rows corresponding to genes and columns corresponding to group memberships.}

\item{I}{Number of maximum iterations.}

\item{thresh}{Convergence criterion.}

\item{.mpmath}{Function depends on mpmath package from python. Pointer for mpmath package.}
}
\description{
Genes are grouped based on the regulating mechanism learned in the first stage using the function, using EMVS.
Then, a group-wise penalty is applied to select genes significantly associated with clinical outcomes while incorporating clinical features using Bayesian variable selection model with extended
NEG priors in \emph{Veronika Ročková & Emmanuel Lesaffre (2014): Incorporating grouping information in Bayesian variable selection with applications in genomics, Bayesian Analysis}
}
\details{
\code{gstr} will take two options "scale" or "1." if \code{gstr} == "scale" then g = 1/N^2 where N = number of genes
}
\examples{
G <- GBM_data2$G
Y <- GBM_data2$Y
C <- GBM_data2$C
R2 <- GBM2_EMVS_res$R2
Zmatrix <- Zmat_builder(R2, G, 0.2, 0.8)
a0 <- 0.1
gstr <- "scale"
mpmath <- setup_mpmath()

.NEG_em(Y, G, C, a0, gstr, Zmatrix,
       I=10, thresh=0.001, .mpmath=mpmath)
}
