% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_multiomics}
\alias{eg_multiomics}
\title{Example multiOmics result}
\format{
An object of class \code{list} of length 2.
}
\usage{
eg_multiomics
}
\description{
Example results of \code{BayesMultiomics} and \code{BayesMultiomicsPriorSensitivity} from vignette("BayesMultiomics")
\itemize{
\item eg_multiomics
\item eg_multiomics_cv
\item eg_multiomics_sensitivity
}
}
\keyword{datasets}
