% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMultiomics.R
\name{plot.BayesMultiomics.CV}
\alias{plot.BayesMultiomics.CV}
\title{plot.BayesMultiomics.CV}
\usage{
\method{plot}{BayesMultiomics.CV}(BayesMultiomics.CV_mod)
}
\arguments{
\item{BayesMultiomics.CV_mod}{BayesMultiomics.CV model object}
}
\description{
Summary box plot of BayesMultiomics.CV model object
}
