% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMultiomics.R
\name{plot.BayesMultiomicsPriorSensitivity}
\alias{plot.BayesMultiomicsPriorSensitivity}
\title{plot.BayesMultiomicsPriorSensitivity}
\usage{
\method{plot}{BayesMultiomicsPriorSensitivity}(BayesMultiomicsPriorSensitivity_obj)
}
\arguments{
\item{BayesMultiomicsPriorSensitivity_obj}{BayesMultiomicsPriorSensitivity object}
}
\description{
plot function for BayesMultiomicsPriorSensitivity object
}
