% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSDIM.R
\name{SSDIM}
\alias{SSDIM}
\title{Second Stage Data Integration Model (SSDIM)}
\usage{
SSDIM(
  Y,
  G,
  C,
  a0,
  gstr,
  Zmatrix,
  I = 10,
  thresh = 0.001,
  .mpmath = setup_mpmath()
)
}
\arguments{
\item{Y}{clinical outcome}

\item{G}{gene expression level}

\item{C}{clinical features}

\item{a0}{size sparsity}

\item{gstr}{prior, two options: "scale" or "1"}

\item{Zmatrix}{Loading matrix, a matrix with rows corresponding to genes and columns corresponding to group memberships.}

\item{I}{number of maximum iterations}

\item{thresh}{convergence criterion}

\item{.mpmath}{function depends on mpmath package from python. pointer for mpmath package}
}
\description{
SSDIM outputs a dataframe of coefficients of a linear model that identifies selected biomarkers.
The linear model is fit on the clinical outcomes \code{Y} using the clinical features \code{C}, gene expression \code{G} and Gene grouping \code{Zmatrix}.

Genes are grouped based on the regulating mechanism learned in the first stage using the function, using EMVS.
Then, a group-wise penalty is applied to select genes significantly associated with clinical outcomes while incorporating clinical features using Bayesian variable selection model with extended
NEG priors in \emph{Veronika Ročková & Emmanuel Lesaffre (2014): Incorporating grouping information in Bayesian variable selection with applications in genomics, Bayesian Analysis}.
}
\details{
\code{gstr} will take two options "scale" or "1." if \code{gstr} == "scale" then g = 1/N^2 where N = number of genes
}
\examples{
G <- GBM_data2$G
Y <- GBM_data2$Y
C <- GBM_data2$C
R2 <- GBM2_EMVS_res$R2
Zmatrix <- Zmat_builder(R2, G, 0.2, 0.8)
mpmath <- setup_mpmath()
a0 <- 0.1
gstr <- "scale"

SSDIM(Y, G, C, a0, gstr, Zmatrix,
      I=10, thresh=0.001, .mpmath=mpmath)
}
