% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMultiomics.R
\name{summary.BayesMultiomicsPriorSensitivity}
\alias{summary.BayesMultiomicsPriorSensitivity}
\title{summary.BayesMultiomicsPriorSensitivity}
\usage{
\method{summary}{BayesMultiomicsPriorSensitivity}(BayesMultiomicsPriorSensitivity_obj)
}
\arguments{
\item{BayesMultiomicsPriorSensitivity_obj}{BayesMultiomicsPriorSensitivity object}
}
\description{
summary method for BayesMultiomicsPriorSensitivity object
}
