% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDIM.R
\name{Zmat_builder}
\alias{Zmat_builder}
\title{Z-matrix builder}
\usage{
Zmat_builder(R2, G, lower_R2 = 0.2, upper_R2 = 0.8)
}
\arguments{
\item{.R2}{R-squared from EMVS result}

\item{.G}{gene expression level}

\item{.lower_R2}{lower limit for R2 threshold, default = 0.2}

\item{.upper_R2}{upper limit for R2 threshold, default = 0.8}
}
\description{
Builds a loading matrix Z that contains gene grouping information based on DNA methylation.
There are 3 groups 1) genes that are modulated only by DNA methylation (M effect),
2) genes that are modulated by characteristics other than DNA methylation (M^c effect);
3) genes that are modulated by both DNA methylation and characteristics other than DNA methylation (M + M^c)

genes with R2 smaller than lower_R2 are put
into M^C effect group, larger than uppder_R2 are put into M effect group and the rest are put into M + M
effect group
}
\examples{
G <- GBM_data2$G
R2 <- GBM2_EMVS_res$R2


Zmat_builder(R2, G, 0.2, 0.8)

}
